last version before code revamp
- changed the name system to have 1 file per player instead of 1 file for all names for optimization purposes
- made the name tag command compatible with offline players

the current way offline players are implemented is extremely laggy when a name tag is set for the first time for an offline player. i plan on optimizing this in the next commit, which would also likely be my last commit on this version of the plugin. after that commit i will most likely revamp the plugin to follow java coding paradigms which should also optimize the plugin in general.