finalized compatibility for offline players
- made it impossible to check an offline player's nametag unless the player is already cached to avoid lag spikes
- fixed an issue where the config command accepted negative integers
- changed the way the nametag is stored for offlineplayer/player parity
- fixed an issue where backups were not stored as yaml files

the way i was initially making this possible was extremely confusing and poorly thought through, so i reworked it a little bit. this was meant to be the final commit before the revamp but i've decided to work on one last feature: extending the name tag visibility distance to the same as a player's nameplate. this is going to take a while.